Thread-safe callbacks for RunnerThreadCallback
Use wx.CallAfter to let wx run the DoseWatcher._runner callbacks that
updates the dose state (both the end/after callback and the exception
callback) instead of calling them directly from the runner thread. It
had never crashed here due to the 50ms sleeps, but that was probably
contingent on the tested environment. Changing the runner pre-spawn
delay to 10ms used to be highly unstable, often crashing with:

  Gdk:ERROR:gdkregion-generic.c:337:miSetExtents:
   assertion failed: (pExtents->y1 < pExtents->y2)
  Aborted (core dumped)