Avoid watchdog issues 117-233 on Linux (inotify)
Watchdog issues 117 and 233 are about a KeyError in
Inotify.read_events that is raised when inotify events are e.g.

  IN_CREATE, IN_DELETE, IN_CREATE, IN_DELETE,
  IN_DELETE_SELF, IN_IGNORE, IN_DELETE_SELF, IN_IGNORE

The first IN_IGNORE removes the wd/path cached for a watcher added by
watchdog due to the second IN_CREATE, so it raises a KeyError when
any event appears with that wd

The fix is monkeypatching the Inotify class to store the last wd;
the patch avoids removing the wd for a path unless they match.

Probably this issue wasn't triggered before in Dose v1.1.0 due to the
50ms delay in the watcher thread