Fix a bug related to fcntl.ioctl in PyCharm/Linux
When running Dose inside PyCharm in a Linux environment, fnctl.ioctl
can't determine the window size (see the method
TerminalSize#from_io_control in terminal.py). As a result of this, the
application will crash even that retrieving this size isn't required at
all. In order to allow Dose to run in this environment, this fix catches
the exact OSError raised internally by the affected commands and
silently ignore them with a brief warning message.